requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
